refactor: add new service for extract info token
